Moving FindWorld() methods to indexer.
